+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+### Added
+- Added `smallerSidebarHeadings` and `widerSidebar` theme config settings that can be enabled to improve class reference sidebar menu headings. [#24](https://github.com/pixelandtonic/vuepress-theme-craftdocs/issues/24)
+
 ### Changed
 - Updated package requirements to resolve security issues.